Implement GeoJSON in Dataset
