--- cartoframes/carto_vl/source/source_types.py
@@ -1,6 +1,0 @@
-from __future__ import absolute_import
-
-
-class SourceTypes():
-    GeoJSON = 'GeoJSON'
-    Query = 'Query'
